% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_homd_data.R
\name{download_homd_data}
\alias{download_homd_data}
\title{Download microbial species data from the Human Oral Microbiome Database (HOMD)
Fourth Function
need to finish
need to somehow implement that its only for substance use}
\usage{
download_homd_data(genus = NULL)
}
\arguments{
\item{genus}{Optional filter by genus (e.g., "Prevotella").}
}
\value{
Data frame with HOMD species data.
}
\description{
Download microbial species data from the Human Oral Microbiome Database (HOMD)
Fourth Function
need to finish
need to somehow implement that its only for substance use
}
