% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_megan.R
\name{parse_megan}
\alias{parse_megan}
\title{Parse MEGAN output for taxonomic data
fifth function
need to double check}
\usage{
parse_megan(megan_file)
}
\arguments{
\item{megan_file}{Path to MEGAN output file.
dont have MEGAN file yet}
}
\value{
A data frame with parsed taxonomic data.
}
\description{
Parse MEGAN output for taxonomic data
fifth function
need to double check
}
