% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_substance.R
\name{plot_species_substance}
\alias{plot_species_substance}
\title{Plot species distribution by substance type
need to finish
need to find substance use only data first}
\usage{
plot_species_substance(data)
}
\arguments{
\item{data}{Microbiome data frame.}
}
\value{
A ggplot object showing species distribution by substance type.
}
\description{
Plot species distribution by substance type
need to finish
need to find substance use only data first
}
