% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run a BLAST search on oral microbial sequences
third function
need to work on still, to get BLAST results correctly}
\usage{
run_blast(sequences)
}
\arguments{
\item{sequences}{A character vector of sequences.}
}
\value{
A data frame with sample BLAST results.
}
\description{
Run a BLAST search on oral microbial sequences
third function
need to work on still, to get BLAST results correctly
}
