% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example Dataset 1: Microbial Taxonomic Data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{Taxon}{Taxonomic group names (e.g., Taxon1, Taxon2, etc.).}
\item{Abundance}{Simulated microbial abundance values.}
\item{Group}{Group category: "Control" or "Substance Users".}
}
}
\usage{
sample_data
}
\description{
Simulated dataset of microbial taxa and their abundances in Control and Substance User groups.
}
\keyword{datasets}
