% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_table.R
\name{species_table}
\alias{species_table}
\title{Generate a species table by substance use type
this need to be finished
need to first find data that belongs to substabce use only}
\usage{
species_table(data)
}
\arguments{
\item{data}{Microbiome data frame.}
}
\value{
Table of species by substance use.
}
\description{
Generate a species table by substance use type
this need to be finished
need to first find data that belongs to substabce use only
}
