% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_taxonomy.R
\name{visualize_taxonomy}
\alias{visualize_taxonomy}
\title{Visualize Taxonomy Profile by Substance Use Type}
\usage{
visualize_taxonomy(taxonomy_data, interactive = FALSE)
}
\arguments{
\item{taxonomy_data}{A data frame containing taxonomic data.
Must include columns: "SubstanceUseType", "Species", and "Abundance".}

\item{interactive}{Logical. If \code{TRUE}, creates an interactive plot using \code{plotly}. Default is \code{FALSE}.}
}
\value{
A \code{ggplot2} object or a \code{plotly} object.
}
\description{
This function creates a bar plot or an interactive plot of taxonomic profiles by substance use type.
}
\details{
This function summarizes taxonomic profiles by substance use type and visualizes them as a bar plot.
}
