% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sequences.R
\name{clean_sequences}
\alias{clean_sequences}
\title{Clean and format sequences for BLAST input
Second Function
need to finish}
\usage{
clean_sequences(raw_data)
}
\arguments{
\item{raw_data}{A data frame with raw sequence data.}
}
\value{
A data frame with cleaned sequences.
}
\description{
Clean and format sequences for BLAST input
Second Function
need to finish
}
