% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare Microbial Communities Between Groups}
\usage{
compare_groups(data, group_col, method = "bray", test_type = "permanova")
}
\arguments{
\item{data}{A data frame containing microbial abundance data.}

\item{group_col}{The name of the column indicating group membership (e.g., "Group").}

\item{method}{The distance method for community comparison (e.g., "bray", "jaccard"). Defaults to "bray".}

\item{test_type}{The statistical test to perform: "permanova" or "ttest".}
}
\value{
A data frame with statistical results.
}
\description{
This function compares microbial communities between two groups using PERMANOVA or t-tests.
}
