% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{Comparing oral microbial communities between substance users and non-users
doesnt work
need to finish}
\usage{
compare_groups(data)
}
\arguments{
\item{data}{Microbiome data with a column for SubstanceUseType.}
}
\value{
A data frame summarizing species differences between groups.
}
\description{
Comparing oral microbial communities between substance users and non-users
doesnt work
need to finish
}
