% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare microbial communities between users and non-users
need to finish
need to somehow demostrate and find data that belongs to substance users and non-users and differentiate them}
\usage{
compare_groups(data)
}
\arguments{
\item{data}{Microbiome data with user group info.}
}
\value{
Summary statistics of microbial differences.
}
\description{
Compare microbial communities between users and non-users
need to finish
need to somehow demostrate and find data that belongs to substance users and non-users and differentiate them
}
