% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_substances.R
\name{compare_substances}
\alias{compare_substances}
\title{Compare microbial communities between substance user types
need to finish
need help with finding the substance use only data}
\usage{
compare_substances(data)
}
\arguments{
\item{data}{Microbiome data with user type info.}
}
\value{
Data frame of comparative statistics.
}
\description{
Compare microbial communities between substance user types
need to finish
need help with finding the substance use only data
}
