% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_homd_data.R
\name{download_homd_data}
\alias{download_homd_data}
\title{Download microbial species data from the Human Oral Microbiome Database (HOMD)
4th function
still working on
this is not working and the test doesnt work}
\usage{
download_homd_data(genus = NULL)
}
\arguments{
\item{genus}{Optional filter by genus (e.g., "Prevotella").}
}
\value{
A data frame with HOMD species data.
}
\description{
Download microbial species data from the Human Oral Microbiome Database (HOMD)
4th function
still working on
this is not working and the test doesnt work
}
