% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Oral microbiome data from a file (sample included in package)
first function}
\usage{
load_data(file_path)

load_data(file_path)
}
\arguments{
\item{file_path}{A string containing the path to the data file.}
}
\value{
A data frame with loaded oral microbiome data.

A preprocessed data frame.
}
\description{
This function loads and preprocesses microbiome data for analysis.
}
