% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_megan.R
\name{parse_megan}
\alias{parse_megan}
\title{Parse MEGAN output for taxonomic data
Fifth function
need to finish
need to double check that MEGAN works
need to double check that MEGAN file is correct}
\usage{
parse_megan(megan_file)
}
\arguments{
\item{megan_file}{Path to MEGAN output file.}
}
\value{
Data frame with taxonomic data.
}
\description{
Parse MEGAN output for taxonomic data
Fifth function
need to finish
need to double check that MEGAN works
need to double check that MEGAN file is correct
}
