% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run a BLAST search on microbial sequences
Third Function
need to finish}
\usage{
run_blast(sequences)
}
\arguments{
\item{sequences}{A character vector of sequences in FASTA format.}
}
\value{
Data frame with BLAST results.
}
\description{
Run a BLAST search on microbial sequences
Third Function
need to finish
}
