% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_table.R
\name{species_table}
\alias{species_table}
\title{Generates a summary species table by substance use type
this need to be finished
need to first find data that belongs to substance use only}
\usage{
species_table(data)
}
\arguments{
\item{data}{Microbiome data with columns for SubstanceUseType and Species.}
}
\value{
A data frame with species counts by substance use type.
}
\description{
Generates a summary species table by substance use type
this need to be finished
need to first find data that belongs to substance use only
}
