% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_taxonomy.R
\name{visualize_taxonomy}
\alias{visualize_taxonomy}
\title{Visualize taxonomy profile by substance type
need to finish
need to find substance use only data}
\usage{
visualize_taxonomy(taxonomy_data)
}
\arguments{
\item{taxonomy_data}{Data frame with taxonomic profiles.}
}
\value{
A ggplot object for taxonomy visualization.
}
\description{
Visualize taxonomy profile by substance type
need to finish
need to find substance use only data
}
