% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_taxonomy.R
\name{visualize_taxonomy}
\alias{visualize_taxonomy}
\title{Visualize taxonomy profile by substance type
need to finish
need to find substance use only data}
\usage{
visualize_taxonomy(tax_data)

visualize_taxonomy(tax_data)
}
\arguments{
\item{tax_data}{A data frame containing taxonomic data.}

\item{taxonomy_data}{Data frame with taxonomic profiles.}
}
\value{
A ggplot object showing taxonomy profile.

A ggplot2 object.
}
\description{
This function creates a bar plot of taxonomic profiles.
}
